+++ NEWS.md
@@ -4,14 +4,14 @@
 
 * BUGFIX: Commands were not always including authentication
   information in every request.
+* IMPROVEMENT: Ruby version compatibility: Now works with Ruby < 1.9.2.
 * NEW: Separate API and CLI help examples: There are now two formats for help
   examples. The new CLI format shows help text as a standard razor-client
   command. CLI is used by default. The API format is the same as before,
   and will show examples in JSON format.
-* IMPROVEMENT: Ruby version compatibility: 0.15.1 would not install on Ruby < 1.9.2.
+* NEW: The `events` collection is new and has a special client-side display.
 * NEW: razor-client now has an 'insecure' flag to ignore SSL verification 
   errors.
-* NEW: The `events` collection is new and has a special client-side display.
 * IMPROVEMENT: Argument types were previously not very context-aware. Now,
   for example, names can include the '=' character.
 * BUGFIX: A reasonable error will be thrown if help is requested but does not exist.